% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaptiveStudy.R
\name{AdaptiveStudy}
\alias{AdaptiveStudy}
\title{AdaptiveStudy}
\usage{
AdaptiveStudy(StudyObj)
}
\arguments{
\item{StudyObj}{An initialized FAIRsimulator \code{study} object.}
}
\value{
A FAIRsimulator \code{study} object complete with all aspects generated during the simulations.
}
\description{
Function that runs a FAIRsimulator simulation.
}
\examples{
\dontrun{
StudyObj<-AdaptiveStudy(StudyObj)
}
}

