% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{AddCohortEvent}
\alias{AddCohortEvent}
\title{AddCohortEvent}
\usage{
AddCohortEvent(StudyObj)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object}
}
\value{
A FAIRsimulator \code{study} object
}
\description{
AddCohortEvent
}

