% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{AddEffect}
\alias{AddEffect}
\title{AddEffect}
\usage{
AddEffect(y, StudyObj, Subject, time, effecttime = 6 * 30, cumeffect = NULL)
}
\arguments{
\item{y}{The HAZ value to add a treatment effect to}

\item{StudyObj}{A FAIRsimulator \code{study} object}

\item{Subject}{A FAIRsimulator \code{individual} object.}

\item{time}{The time of the HAZ observation}

\item{effecttime}{The effecttime}

\item{cumeffect}{Should the effect be cumulative to previous cohorts.}
}
\description{
Add a treatment effect to a HAZ value
}

