% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{AddNewBirthCohortEvent}
\alias{AddNewBirthCohortEvent}
\title{AddNewBirthCohortEvent}
\usage{
AddNewBirthCohortEvent(StudyObj)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object}
}
\value{
A FAIRsimulator \code{study} object
}
\description{
Function for adding new cohorts at birth based on previous ended birth cohorts
}

