% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{AnalyzeDataEvent}
\alias{AnalyzeDataEvent}
\title{AnalyzeDataEvent}
\usage{
AnalyzeDataEvent(StudyObj)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{cohort} object}
}
\value{
A FAIRsimulator \code{cohort} object
}
\description{
An event to check if the cohort is completed and also set that subjects are completed according to the criteria set.
}
\examples{
\dontrun{}
}

