% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHAZVector.R
\name{calcHAZVector}
\alias{calcHAZVector}
\title{calHAZVector}
\usage{
calcHAZVector(age, basethetas = NULL, covthetas = rep(0,
  length(basethetas)), etas = rep(0, length(basethetas)), BASE, PLMAX, HLKON,
  HLKOFF, BASSL, BP, BASECOV = 0, PLMAXCOV = 0, HLKONCOV = 0,
  HLKOFFCOV = 0, BASSLCOV = 0, BPCOV = 0)
}
\arguments{
\item{age}{A vector of ages (in years) for which to compute the HAZ values.}

\item{basethetas}{A vector of base model thetas (fixed effects).}

\item{covthetas}{A vector with the FFEM, parameter specific covariate effetcs.}

\item{etas}{A vecor of individual specific random effects.}

\item{BASE}{The fixed effects parameter value for the BASE parameter in the FREM  model.}

\item{PLMAX}{The fixed effects parameter value for the PLMAX parameter in the FREM  model.}

\item{HLKON}{The fixed effects parameter value for the HLKON parameter in the FREM  model.}

\item{HLKOFF}{The fixed effects parameter value for the HLKOFF parameter in the FREM  model.}

\item{BASSL}{The fixed effects parameter value for the BASSL parameter in the FREM  model.}

\item{BP}{The fixed effects parameter value for the BP parameter in the FREM  model.}

\item{BASECOV}{The FFEM covariate effect for the BASE parameter in the FREM  model.}

\item{PLMAXCOV}{The FFEM covariate effect for the PLMAX parameter in the FREM  model.}

\item{HLKONCOV}{The FFEM covariate effect for the HLKON parameter in the FREM  model.}

\item{HLKOFFCOV}{The FFEM covariate effect for the HLKOFF parameter in the FREM  model.}

\item{BASSLCOV}{The FFEM covariate effect for the BASSL parameter in the FREM  model.}

\item{BPCOV}{The FFEM covariate effect for the BP parameter in the FREM  model.}
}
\value{
A vector of HAZ values as long as the vector of \code{age}.
}
\description{
Computes a vector of individual HAZ values based on the FREM model.
}
\details{
The basethetas+covthetas+etas and BASE+PLMAX+HLKON+HLKOFF+BASSL+BP+(the corrsponding covariate effects) are mutually exclusive. 
If basethetas is provided, then the former set of arguments will be used.
}
\examples{
\dontrun{
hazValues <- calcHAZVector(age = c(0,30,60,90)/(30*12),basethetas = thbasevector,covthetas = FREMCoeff,etas = IndEtas)
}
}

