% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{CohortCompleted}
\alias{CohortCompleted}
\title{CohortCompleted}
\usage{
CohortCompleted(Cohort, StudyObj)
}
\arguments{
\item{Cohort}{A FAIRsimulator \code{cohort} object.}

\item{StudyObj}{A FAIRsimulator \code{study} object.}
}
\value{
TRUE or FALSE
}
\description{
Function to check if a cohort is completed.
}
\examples{
\dontrun{}
}

