% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{cohorts}
\alias{cohorts}
\title{cohorts}
\usage{
cohorts(obj)
}
\arguments{
\item{obj}{A FAIRsimulator \code{study} object}
}
\value{
The list of cohorts in the \code{study} object.
}
\description{
Extract the list of cohorts from a FAIRsimulator \code{study} object
}
\examples{
\dontrun{
myCohorts <- cohorts(StudyObj)
}
}

