% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStudy.R
\name{createStudy}
\alias{createStudy}
\title{createStudy}
\usage{
createStudy(nCohorts = 3, cohortStartTimes = c(2, 1, 0),
  samplingDesign = list(c(0, 1, 2, 3, 4, 5, 6) * 30, c(0/30, 3, 6) * 30,
  c(0/30, 3, 6) * 30), nSubjects = c(320, 320, 320), dropoutRates = c(0.2/(6
  * 30), 0.2/(6 * 30), 0.2/(6 * 30)), recruitmentAges = list(c(0, 1) * 30,
  c(6, 7) * 30, c(12, 13) * 30),
  RecruitmentFunction = RecruitmentRatefunction, debugLevel = 1)
}
\arguments{
\item{nCohorts}{The number of cohorts.}

\item{cohortStartTimes}{A vector of start times for the different cohorts, in days of study time.}

\item{samplingDesign}{A list of vector with sampling times for each cohort.}

\item{nSubjects}{A vector of number of subjects in each cohort.}

\item{dropoutRates}{A vector of dropout rates (per day) for each of the cohorts}

\item{recruitmentAges}{A list of age ranges in which subjects are recruited to each cohort. Each list component should be a vector with the min and max age.}

\item{debugLevel}{The debug level. 4=Extreme output, #3=Print everything important, 2=Print events, 1=Sparse print, 0=Print nothing}
}
\value{
A FAIRsimulator \code{study} object ready to be executed.
}
\description{
Create an initial study object that represents the study state before it is executed.
}

