% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{DebugPrint}
\alias{DebugPrint}
\title{DebugPrint}
\usage{
DebugPrint(str, iLevel, StudyObj)
}
\arguments{
\item{str}{A string.}

\item{iLevel}{The level of debug messages.}

\item{StudyObj}{A FAIRsimulator \code{study} object}
}
\value{
A debug message.
}
\description{
Function to print debug information
}
\examples{
\dontrun{

}
}

