% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{DropoutEvent}
\alias{DropoutEvent}
\title{DropoutEvent}
\usage{
DropoutEvent(StudyObj)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object}
}
\value{
A FAIRsimulator \code{study} object
}
\description{
Event that check if subject have dropped out
}

