% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{DropoutTime}
\alias{DropoutTime}
\title{DropoutTime}
\usage{
DropoutTime(lambda, randnum = NULL)
}
\arguments{
\item{lambda}{The lambda value}

\item{randnum}{An optional random number. Will be generated by the fuction if NULL.}
}
\value{
The dropout day.
}
\description{
Constant hazard dropout function
}

