% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInclusionData.R
\name{extractInclusionData}
\alias{extractInclusionData}
\title{extractInclusionData}
\usage{
extractInclusionData(StudyObj)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object}
}
\value{
A \code{data.frame} with columns: StudyDay, Recruited, Dropout, Completed, Remaining, CohortDay, CohortID, CycleNum, CohortStartTime, Cycle and Cohort.
}
\description{
Extracts statistics on recruitment, dropout and completion of subjetcs from each cohort.
}
\examples{
\dontrun{
indInfo <- extractInclusionData(StudyObj) 
}
}

