% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{futilityFunction}
\alias{futilityFunction}
\title{futilityFunction}
\usage{
futilityFunction(probs, Cohort, StudyObj,
  minSubj = StudyObj$StudyDesignSettings$MinimumNumberofSubjects)
}
\arguments{
\item{probs}{The current randomization probabilities after having taken the minimum randomisation probabilities into account.}

\item{Cohort}{A FAIRsimulator \code{cohort} object}

\item{StudyObj}{A FAIRsimulator \code{study} object}

\item{minSubj}{The cutoff for the number of subjects expected to be randomised to a treatment.}
}
\value{
A vector of updated probabilities.
}
\description{
Function to determine if treatments are futile.
}
\details{
The function checks the expected number of subjects to be randomised to a particular treatment given the probabilities in \code{probs}. Any treatment that has an expected number of subjects 
lower than or equal to \code{minSubj} will have their randomisation probabilitiy set to 0. The remaining probabilities will be adjusted to add up to 1.
}

