% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{GetAgesAtRandomization}
\alias{GetAgesAtRandomization}
\title{GetAgesAtRandomization}
\usage{
GetAgesAtRandomization(NumRecruited, Cohort, StudyObj)
}
\arguments{
\item{NumRecruited}{The number of ages to return}

\item{Cohort}{A FAIRsimulator \code{cohort} object}

\item{StudyObj}{A FAIRsimulator \code{study} object}
}
\value{
A vector of uniform ages
}
\description{
Returns the uniform age at randomization given an age range
}

