% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllSubjectData.R
\name{getAllSubjectData}
\alias{getAllSubjectData}
\title{getAllSubjectData}
\usage{
getAllSubjectData(StudyObj, ...)
}
\arguments{
\item{StudyObj}{A FAIRsimulator study object}

\item{...}{Parameters passed to \code{\link{getSubjectItems}}}
}
\value{
A data frame
}
\description{
Extracts data framm all subjects in all cohorts in a FAIRsimulator study object
}
\examples{
\dontrun{
tmp <- getAllSubjectData(StudyObj)
}
}
\seealso{
\code{\link{getSubjectItems}}, \code{\link{getItemsFromSubjects}}
}

