% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{getCohortAgeRange}
\alias{getCohortAgeRange}
\title{getCohortAgeRange}
\usage{
getCohortAgeRange(StudyObj, cohortAgeNames = c("0-6 months", "6-12 months",
  "12-18 months"))
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object}

\item{cohortAgeNames}{An optional vecto of labels for the age groups}
}
\value{
A data.frame with columns CohortName, AgeRange and CohortAge
}
\description{
Creates a lookup data .frame with CohortName, Recruitment age range and a Label for the age group.
}
\examples{
\dontrun{
getCohortAgeRange(StudyObj)
}
}

