% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCohortDetails.R
\name{getCohortDetails}
\alias{getCohortDetails}
\title{getCohortDetails}
\usage{
getCohortDetails(StudyObj)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object.}
}
\value{
A data.frame with values
}
\description{
Extract scalar statistics for the cohorts in a FAIRsimulator \code{study} object.
}
\examples{
\dontrun{
getCohortDetails(StudyObj)
}
}

