% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRsimulatorFunctions.R
\name{getCohortNames}
\alias{getCohortNames}
\title{getCohortNames}
\usage{
getCohortNames(StudyObj)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object.}
}
\value{
A character vector
}
\description{
Extracts the names of the cohorts in the CohortList is a FAIRsimulator \code{study} object.
}
\details{
This is an internal function that extracts the values of the \code{Name} slot in the \code{CohortList} list of a FAIRsimulator \code{study} object. 
It is typically used to set the names of the \code{CohortList} components, so that they can be extracted using \code{names(CohortList)} and referred to by name, e.g. 
\code{CohortList[[cohortName]]}.
}
\examples{
\dontrun{
names(StudyObj$CohortList) <- getCohortNames(StudyObj)
}
}

