% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRsimulatorFunctions.R
\name{getItemsFromSubjects}
\alias{getItemsFromSubjects}
\title{getItemsFromSubjects}
\usage{
getItemsFromSubjects(cohortObj, ...)
}
\arguments{
\item{cohortObj}{A FAIRsimulator cohort object}

\item{...}{Parameters passed to \code{\link{getSubjectItems}}}
}
\value{
A data frame
}
\description{
Extracts data from all subjects in a FAIRsimulator cohort object
}
\examples{
\\donrun{
tmp <- getItemsFromSubjects(StudyObj[["CohortList"]][[1]])
}
}
\seealso{
\code{\link{getSubjectItems}}
}

