% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{getMultiProbList}
\alias{getMultiProbList}
\title{getMultiProbList}
\usage{
getMultiProbList(multiStudObj, ncores = 1, ...)
}
\arguments{
\item{multiStudObj}{A list of FAIRsimulator \code{study} objects}

\item{ncores}{The number of cores to use.}

\item{...}{Arguments to \code{getProbData}, in particular \code{strProb}}
}
\value{
A data frame with probabilities. Includes a column with Iteration number.
}
\description{
Extracts the probability lists from a list of FAIRsimulator \code{study} objects. It will optionally be done in parallel.
}
\examples{
\dontrun{
getMultiProbList(multiStudObj,ncores=2,strProb="UnWeightedRandomizationProbabilities") 
}
}

