% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{GetNewRandomizationProbabilities}
\alias{GetNewRandomizationProbabilities}
\title{GetNewRandomizationProbabilities}
\usage{
GetNewRandomizationProbabilities(trtcoeff, trtse, iNumPosteriorSamples,
  randsam = NULL)
}
\arguments{
\item{trtcoeff}{Treatment coefficients}

\item{trtse}{Standard errors of the treatment coefficients}

\item{iNumPosteriorSamples}{Number of samples to use for computing the posterior probability}

\item{randsam}{An object}
}
\value{
A vector of the same length as the \code{trtcoeff} with the updated posterior probabilities for each treatment to be the best treatment.
}
\description{
Function for calculating best treatment probabilitites based on sampling from estimated treatments effect with SEs. 
Can also calculate best treatment based on random_sampling (randsam) variable with dim of NumTreatments x iNumPosteriorSamples.
}
\examples{
\dontrun{}
}

