% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{getProbData}
\alias{getProbData}
\title{getProbData}
\usage{
getProbData(StudyObj, strProb = "UpdateProbabilities", ...)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object}

\item{strProb}{A string indicating the probabilities to extract. Possible values are "UnWeightedRandomizationProbabilities" and "RandomizationProbabilities". The former 
are the probabilities after adjustment for minimum allocation probabilities and the latter are the unadjusted probabilities.}

\item{...}{Other parameters passed to \code{getCohortAgeRange}}
}
\value{
A data.frame with columns CohortName (cohort name), Prob (probability),  TreatmentName, CohortAge (cohort age group label) and  RandStudyTime 
(the time the randomisation probabilities were set).
}
\description{
Extracts the randomization probabilities from all cohorts over time.
}
\examples{
\dontrun{
probData <- getProbData(StudyObj)
}
}

