% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRandomizationProbabilities.R
\name{getRandomizationProbabilities}
\alias{getRandomizationProbabilities}
\title{getRandomizationProbabilities}
\usage{
getRandomizationProbabilities(StudyObj)
}
\arguments{
\item{StudyObj}{A FAIRsimulator study object}
}
\value{
A data frame
}
\description{
Extract randomization probabilities over time/treatment
}
\examples{
\dontrun{
}
}

