% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{GetRecruitedSubjects}
\alias{GetRecruitedSubjects}
\title{GetRecruitedSubjects}
\usage{
GetRecruitedSubjects(StudyObj, Cohort)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object}

\item{Cohort}{A FAIRsimulator \code{cohort} object}
}
\value{
The number of subjects
}
\description{
Get the number of recruited subjects at current time.
}

