% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{GetStartNumIndex}
\alias{GetStartNumIndex}
\title{GetStartNumIndex}
\usage{
GetStartNumIndex(StudyObj, StartNum)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object}

\item{StartNum}{The cohort start number}
}
\value{
The cohort list index
}
\description{
Return the cohort list index of a certain cohort start number
}

