% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{GetSubject}
\alias{GetSubject}
\title{GetSubject}
\usage{
GetSubject(ID, TRTIndex, AgeAtRand, Cohort, StudyObj)
}
\arguments{
\item{ID}{The subject identification number}

\item{TRTIndex}{The treatment index}

\item{AgeAtRand}{The age at randomization}

\item{Cohort}{A FAIRsimulator \code{cohort} object}

\item{StudyObj}{A FAIRsimulator \code{study} object}
}
\value{
A FAIRsimulator \code{individual} object.
}
\description{
Creates a FAIRsimulator \code{individual} object.
}

