% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRsimulatorFunctions.R
\name{getSubjectItems}
\alias{getSubjectItems}
\title{getSubjectItems}
\usage{
getSubjectItems(subjectObj, scalarItems = c("StudyID", "AgeAtRand",
  "DateAtRand", "RandStudyTime", "RandCohortTime", "CurrentAge",
  "CurrentCohortTime", "TreatmentIndex", "Treatment", "TreatmentEff", "RandNum",
  "DropoutStudyTime", "DropoutCohortTime"), covariates = "Covariates",
  longitudinalItems = c("SampleAge", "CohortSampleTime", "StudySampleTime",
  "Data"))
}
\arguments{
\item{subjectObj}{A FAIRsimulator individual object}

\item{scalarItems}{The scalar items to extract from the individual object. Set to \code{NULL} if no scalar items are to be extracted.}

\item{covariates}{Name of the item that holds the subject specific vector of covariates. Set to \code{NULL} if covariates should not be extracted.}

\item{longitudinalItems}{Name of the longitudinal items to be extracted. Set to \code{NULL} if no no longitudina items are to be extracted.}
}
\value{
A data frame with nrows=length of the longitudinal items. The scalar and covariate items are replicated to match the longitudinal items.
}
\description{
Extracts all subject level data as a data.frame from a FAIRsimulator individual object
}
\examples{
\dontrun{
getSubjectItems(StudyObj[["CohortList"]][[1]][["SubjectList"]][[1]])
#' }
}

