% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubjectLevelItems.R
\name{getSubjectLevelItems}
\alias{getSubjectLevelItems}
\title{getSubjectLevelItems}
\usage{
getSubjectLevelItems(studyObj, items = "AgeAtRand")
}
\arguments{
\item{studyObj}{A FAIRsimulator \code{study} object}

\item{items}{A vector of the scalar subject level items to extract for each subject in each cohort.}
}
\value{
A data.frame with the columns: cohort name, SubjID and each of the entries in \code{item}.
}
\description{
Extract scalar subject level items from all subjects in all cohorts in a FAIRsimulator \code{study} object
}
\examples{
\dontrun{
tmp <- getSubjectLevelItem(StudyObj,item=c("AgeAtRand","RandCohortTime"))
}
}

