% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{GetSubjects}
\alias{GetSubjects}
\title{GetSubjects}
\usage{
GetSubjects(NumRecruited, TreatmentIndex, RandAges, Cohort, StudyObj)
}
\arguments{
\item{NumRecruited}{The number of subjects to recruit.}

\item{TreatmentIndex}{The treatment index}

\item{RandAges}{The randomization ages}

\item{Cohort}{A FAIRsimulator \code{cohort} object}

\item{StudyObj}{A FAIRsimulator \code{study} object}
}
\value{
A list of FAIRsimulator \code{individual} objects.
}
\description{
Create a list of subjects
}

