% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{GetTreatmentRandomizations}
\alias{GetTreatmentRandomizations}
\title{GetTreatmentRandomizations}
\usage{
GetTreatmentRandomizations(Num, Cohort, StudyObj)
}
\arguments{
\item{Num}{The treatment randomization probabilities.}

\item{Cohort}{A FAIRsimulator \code{cohort} object}

\item{StudyObj}{A FAIRsimulator \code{study} object}
}
\value{
The treatment index, i.e. the treatment the subjects are randomized to.
}
\description{
Get the treatment randomizations
}

