% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotColours.R
\name{ggplotColours}
\alias{ggplotColours}
\title{ggplotColours}
\usage{
ggplotColours(n = 6, h = c(0, 360) + 15)
}
\arguments{
\item{n}{number of colors needed.}

\item{h}{an argument}
}
\value{
a vector of colors
}
\description{
Extracts the standard ggplot colors, given the number of colors needed.
}
\examples{
ggplotColours(4)
}

