% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listmap.R
\name{\%listmap\%}
\alias{\%listmap\%}
\title{listmap}
\usage{
x \%listmap\% n
}
\arguments{
\item{x}{The list to extract elements from}

\item{n}{The name of the element}
}
\value{
The element
}
\description{
An operator that extracts elements from a lisy
}
\examples{
\dontrun{
mySubjects \%listmap\% item
}
}

