% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRsimulatorFunctions.R
\name{\%map\%}
\alias{\%map\%}
\title{map}
\usage{
x \%map\% n
}
\arguments{
\item{x}{The list to extract elements from}

\item{n}{The name of the element}
}
\value{
The element
}
\description{
An operator that extracts elements from a lisy
}
\examples{
\dontrun{
mySubjects \%map\% item
}
}

