% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{InitEvent}
\alias{InitEvent}
\title{InitEvent}
\usage{
InitEvent(StudyObj)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object}
}
\value{
A FAIRsimulator \code{study} object
}
\description{
Initialization event function. Reads in the necessary FREM stuff for the simulations.
}

