% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{InterimAnalyzesTime}
\alias{InterimAnalyzesTime}
\title{InterimAnalyzesTime}
\usage{
InterimAnalyzesTime(Cohort, StudyObj)
}
\arguments{
\item{Cohort}{A FAIRsimulator \code{cohort} object}

\item{StudyObj}{A FAIRsimulator \code{study} object}
}
\value{
TRUE or FALSE
}
\description{
Check if it istime to do an interim analysis for a specific cohrt
}
\examples{
\dontrun{}
}

