% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{MoveCompletedSubjects}
\alias{MoveCompletedSubjects}
\title{MoveCompletedSubjects}
\usage{
MoveCompletedSubjects(StudyObj)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object}
}
\value{
A FAIRsimulator \code{study} object
}
\description{
Event that checks if completed subjects should be moved to new Cohort. Only subjects that are in a cohort that is linked to another cohort is affected.
}

