% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{MoveSubjects}
\alias{MoveSubjects}
\title{MoveSubjects}
\usage{
MoveSubjects(FromCohort, ToCohort, StudyObj)
}
\arguments{
\item{FromCohort}{A FAIRsimulator \code{cohort} object for the cohort to move subjects from.}

\item{ToCohort}{A FAIRsimulator \code{cohort} object for the cohort to move subjects to.}

\item{StudyObj}{A FAIRsimulator \code{study} object}
}
\value{
A FAIRsimulator \code{cohort} object (the ToCohort)
}
\description{
Move subjects from FromCohort to ToCohort which are completed (Status==2), Re-randomize treatments based on ToCohort rand probabilities
}

