% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{NewCohort}
\alias{NewCohort}
\title{NewCohort}
\usage{
NewCohort(StudyObj, CohortNum = NULL)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object}

\item{CohortNum}{The cohort index number to update.}
}
\value{
A FAIRsimulator \code{cohort} object
}
\description{
Create a newFAIRsimulator \code{cohort} object in a FAIRsimulator \code{study} object, or update an existing one.
}
\examples{
\dontrun{}
}

