% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotActiveSubjects.R
\name{plotActiveSubjects}
\alias{plotActiveSubjects}
\title{plotActiveSubjects}
\usage{
plotActiveSubjects(StudyObj)
}
\arguments{
\item{StudyObject}{A FAIRsimulator \code{study} object.}
}
\value{
A ggplot object.
}
\description{
Plot the number of active subjects in each cohort and cycle.
}
\details{
The function returns a graph of the number of subjects in each cohort and cycle. The number is computed as the 
cumulated number of recruted subjects minus the cumulated number subjects that 
have dropped out minus the cumulated number of subjects that have completed the cohort.
}
\examples{
\dontrun{
plotActiveSubjects(StudyObject)
}
}

