% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHAZ.R
\name{plotHAZ}
\alias{plotHAZ}
\title{plotHAZ}
\usage{
plotHAZ(StudyObj, data = NULL)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object.}

\item{data}{A data frame with the data to be plotted. If omitted, the data will be extracted from the study object.}
}
\value{
A \code{ggplot} object
}
\description{
Plots the simulated data in a FAIRsimulator \code{study} object.
}
\examples{
\dontrun{
plotHAZ(StudyObj)
}
}

