% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{plotMultiProb}
\alias{plotMultiProb}
\title{plotMultiProb}
\usage{
plotMultiProb(probDf, ylb = "Randomization probability", pup = 0.95,
  pdo = 0.05)
}
\arguments{
\item{probDf}{A data.frame with probabilities for multiple FAIRsimulator \code{study} objects.}

\item{ylb}{y-axis label for}

\item{pup}{Upper percentile for the prediction interval.}

\item{pdo}{Lower percentile for the prediction interval.}
}
\value{
An invisible list of ggplot objects as long as the number of cohorts.
}
\description{
Plots the updated probabilities over study time, per cohort, with prediction interval.
}
\examples{
\dontrun{
plotMultiProb(probDf)
plotMultiProb(probDfUnweighted,ylb="Unweighted randomization probabilities")
}
}

