% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{plotProbs}
\alias{plotProbs}
\title{plotProbs}
\usage{
plotProbs(StudyObj, ...)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object}

\item{...}{Other parameters passed to \code{getCohortAgeRange}}
}
\value{
An invisible list of ggplot objects, one for each age tretament group.
}
\description{
Plots the randomization probabilities over study time stratified by age treatment group.
}
\examples{
\dontrun{
plotProbs(StudyObj)
}
}

