<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRsimulatorFunctions.R
\name{plotStudyCohorts}
\alias{plotStudyCohorts}
\title{plotStudyCohorts}
\usage{
plotStudyCohorts(StudyObj, cohortlength = 6 * 30)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object.}

\item{cohortlength}{The length (in days) of cohorts that are evolving}
}
\value{
A \code{ggplot} object
}
\description{
Plots the cohort durations/cycles in a FAIRsimulator \code{study} object.
}
\examples{
#' \dontrun{
plotStudyCohorts(StudyObj)
}
}

=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStudyCohorts.R
\name{plotStudyCohorts}
\alias{plotStudyCohorts}
\title{plotStudyCohorts}
\usage{
plotStudyCohorts(StudyObj, cohortlength = 6 * 30, wrapVariable = NULL)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object.}

\item{cohortlength}{The length (in days) of cohorts that are evolving}

\item{wrapVariable}{The name (as a string) of the variable to create facets with, using \code{facet_wrap}.}
}
\value{
A \code{ggplot} object
}
\description{
Plots the cohort durations/cycles in a FAIRsimulator \code{study} object.
}
\examples{
\dontrun{
plotStudyCohorts(StudyObj)
}
}

>>>>>>> a3645c8b8ba3a07bde82b05b7e21d3276b10ba3f
