% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{probTemperation}
\alias{probTemperation}
\title{probTemperation}
\usage{
probTemperation(probs)
}
\arguments{
\item{probs}{The current set of probabilities.}
}
\value{
A set of probabilities
}
\description{
This funciton is called after each update of the probabilities. It is a placeholder for a function that makes the randomization updates less drastic. 
The default function does nothing to teh probabilities.
}
\examples{
\dontrun{
probTemperation(probs) # Default. Will give back the same probabilties.
probTemperation <- function(probs) {
     probs <- sqrt(probs)/sum(sqrt(probs))
     return(probs)
}
probTemperation(probs) # Will now give back the less dramatic square rooted probabilities.
}
}

