% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{RecruitmentEvent}
\alias{RecruitmentEvent}
\title{RecruitmentEvent}
\usage{
RecruitmentEvent(StudyObj)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object}
}
\value{
A FAIRsimulator \code{study} object
}
\description{
Event that recruites subjects to cohort
}

