% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{RecruitmentRatefunction}
\alias{RecruitmentRatefunction}
\title{RecruitmentRatefunction}
\usage{
RecruitmentRatefunction(StudyObj, Cohort)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{cohort} object}

\item{Cohort}{}
}
\value{
The target number of subjects to recruit per day.
}
\description{
Function to give back the recruitment rate from a FAIRsimulator \code{study} object.
}
\examples{
\dontrun{}
}

