% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{runMultiSim}
\alias{runMultiSim}
\title{runMultiSim}
\usage{
runMultiSim(StudyOnjIni, extractProbs = TRUE, iter = 1, ncores = 1,
  strProb = "UnWeightedRandomizationProbabilities", cohortAgeNames = NULL,
  clean = TRUE)
}
\arguments{
\item{StudyOnjIni}{An initial FAIRsimulator \code{study} object created by \code{createStudy}.}

\item{extractProbs}{Whether to also extract randomization probabilities (in parallel).}

\item{iter}{The number of studies to simulate.}

\item{ncores}{The number of cores to use for parallel execution.}

\item{...}{Other parameters. Typically \code{strProb} to specify "UnWeightedRandomizationProbabilities" instead of the default "RandomizationProbabilities".}
}
\value{
A list with a list if the simulated FAIRsimulator \code{study} objects and, optionally, a data.frame with the randomizaion probabilities.
}
\description{
Simulate multiple instances of a study, optionally in a parallel fashion.
}
\examples{
\dontrun{
StudyObjIni <- createStudy(latestTimeForNewBirthCohorts=18*30,studyStopTime = 32*30,
nSubjects = c(320,320,320),
randomizationProbabilities = list(rep(0.20,5),rep(0.20,5),rep(0.20,5)),
#minAllocationProbabilities = list(c(0,rep(0,4)),c(0,rep(0,4)),c(0,rep(0,4))),
minAllocationProbabilities = list(c(0.2,rep(0,4)),c(0.2,rep(0,4)),c(0.2,rep(0,4))),
treatments =list(c("SoC-1","TRT-1","TRT-2","TRT-3","TRT-4"),c("SoC-2","TRT-5","TRT-6","TRT-7","TRT-8"),c("SoC-3","TRT-9","TRT-10","TRT-11","TRT-12")),
effSizes = list(c(0,0.05,0.1,0.15,0.25),c(0,0.05,0.1,0.15,0.25),c(0,0.05,0.1,0.15,0.25)),
Recruitmentfunction=function(...) {return(5000)},
minSubjects = 10)

myMultStud <- runMultiSim(StudyOnjIni,iter=10,ncores=2)
}
}

