% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{SimulateDataEvent}
\alias{SimulateDataEvent}
\title{SimulateDataEvent}
\usage{
SimulateDataEvent(StudyObj)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object}
}
\value{
A FAIRsimulator \code{study} object
}
\description{
Simulate data for all subjects which should have a sample at this date
}

