% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{SimulateHAZ}
\alias{SimulateHAZ}
\title{SimulateHAZ}
\usage{
SimulateHAZ(StudyObj, Subject, time, age)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object}

\item{Subject}{A FAIRsimulator \code{individual} object.}

\item{time}{The time of the new HAZ value}

\item{age}{The age of the subject}
}
\value{
The simulated HAZ value
}
\description{
Simulate HAZ values
}

