% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{StopEvent}
\alias{StopEvent}
\title{StopEvent}
\usage{
StopEvent(StudyObj)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{cohort} object}
}
\value{
The stop criteria for the study.
}
\description{
The stop criteria  for the whole study
}
\examples{
\dontrun{}
}

