% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{StudyIncrementEvent}
\alias{StudyIncrementEvent}
\title{StudyIncrementEvent}
\usage{
StudyIncrementEvent(StudyObj)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object}
}
\value{
A FAIRsimulator \code{study} object
}
\description{
Increments the time in a FAIRsimulator \code{study} object.
}
\examples{
\dontrun{}
}

