% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjects.R
\name{subjects}
\alias{subjects}
\title{subjects}
\usage{
subjects(obj)
}
\arguments{
\item{obj}{A FAIRsimulator \code{cohort} object}
}
\value{
The list of cohorts in the \code{cohort} object.
}
\description{
Extract the list of subjects from a FAIRsimulator \code{cohort} object
}
\examples{
\dontrun{
mySubjects <- subjects(StudyObj)
}
}

