% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{UpdateProbabilities}
\alias{UpdateProbabilities}
\title{UpdateProbabilities}
\usage{
UpdateProbabilities(Cohort, StudyObj, cohortindex = NULL)
}
\arguments{
\item{Cohort}{A FAIRsimulator \code{cohort} object}

\item{StudyObj}{A FAIRsimulator \code{study} object}

\item{cohortindex}{Cohort index.}
}
\value{
A FAIRsimulator \code{cohort} object
}
\description{
Update probabilities based on data analysis and probability of being best
}
\examples{
\dontrun{}
}

