% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{UpdateProbabilitiesEvent}
\alias{UpdateProbabilitiesEvent}
\title{UpdateProbabilitiesEvent}
\usage{
UpdateProbabilitiesEvent(StudyObj)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object}
}
\value{
A FAIRsimulator \code{study} object
}
\description{
Update probabilities of non-connected cohorts.
}

