% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{updateProbs}
\alias{updateProbs}
\title{updateProbs}
\usage{
updateProbs(probs, Cohort, minProb = Cohort$MinAllocationProbabilities)
}
\arguments{
\item{probs}{A vector of probabilities to be adjusted. The vector sum does not have to be 1.}

\item{Cohort}{A FAIRsimulator \code{cohort} object}

\item{minProb}{A vector of the same length as \code{probs} with the minimum values.}
}
\value{
A vector of probabilties in which the minimum probabilties are honored and the remaining probabilities are adjusted so that the vector sum is 1.
}
\description{
Updates a vector of probabilties so that one or more of the probabilities are of at least a certain size.
}
\details{
Typically used to ensure that standard of care always have a certain randomisation probability.
}

