% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{GetCohortData}
\alias{GetCohortData}
\title{GetCohortData}
\usage{
GetCohortData(Cohort, StudyObj)
}
\arguments{
\item{Cohort}{A FAIRsimulator \code{cohort} object}

\item{StudyObj}{A FAIRsimulator \code{study} object}
}
\value{
A \code{data.frame}
}
\description{
Get dataset for a Cohort object
}
\examples{
\dontrun{}
}

