% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRStudy.R
\name{ImputeCovariates}
\alias{ImputeCovariates}
\title{ImputeCovariates}
\usage{
ImputeCovariates(df, StudyObj, strIDVariable = "ID")
}
\arguments{
\item{df}{A \code{data.frame} in which to impute missing covariate values.}

\item{StudyObj}{A FAIRsimulator \code{study} object}

\item{strID}{A string with the name of the columns in \code{df} containing the subject identifiers.}
}
\value{
A \code{data.frame} with all missing covariates imputed.
}
\description{
Impute covariate - simple median imputation per individual (strIDVariable) of data frame (without correlation) of covariates in covariates vector.
}
\examples{
\dontrun{}
}

