% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHAZTreatmentEff.R
\name{plotHAZTreatmentEff}
\alias{plotHAZTreatmentEff}
\title{plotHAZTreatmentEff}
\usage{
plotHAZTreatmentEff(StudyObj)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object.}
}
\value{
A \code{ggplot} object
}
\description{
Plots the simulated data in a FAIRsimulator \code{study} object with treatment effects.
}
\details{
Will produce a graph of individual change from cohort baseline HAZ. Treatment effects are visualised by linear regression lines
for each treatment group.
}
\examples{
\dontrun{
plotHAZTreatmentEff(StudyObj)
}
}

