% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAIRsimulatorFunctions.R
\name{plotStudyCohorts}
\alias{plotStudyCohorts}
\title{plotStudyCohorts}
\usage{
plotStudyCohorts(StudyObj, cohortlength = 6 * 30)
}
\arguments{
\item{StudyObj}{A FAIRsimulator \code{study} object.}

\item{cohortlength}{The length (in days) of cohorts that are evolving}
}
\value{
A \code{ggplot} object
}
\description{
Plots the cohort durations/cycles in a FAIRsimulator \code{study} object.
}
\examples{
#' \dontrun{
plotStudyCohorts(StudyObj)
}
}

